requirements.txt
streamlit
pandas 
yfinance 
ta
matplotlib.pyplot 